break out logic for rendering the default list view or the edit mode view

+++ src/components/List.jsx
@@ -10,20 +10,38 @@ const List = ({name, price, count, addOneMoreToCart, deleteOneMoreFromCart, item
         deleteItemFromOptionList(itemId, price, count);
     }, [deleteItemFromOptionList, itemId, price, count]);
 
-    return ( 
-        <li>
-            { (!editMode) ? 
-            (<> <span name={name}>{name}</span> 
+    let defaultListLayout = null;
+    let editModeLayout = null;
+    
+    if(admin) {
+        defaultListLayout = (
+            <>
+                <span name={name}>{name}</span> 
+                <button onClick={() => setEditMode(true)}>Edit</button> 
+                <button onClick={() => handleDeleteItem()}>Delete</button> 
+            </>
+        )
+    } else {
+        defaultListLayout = (
+            <>
+                <span name={name}>{name}</span>
                 <button onClick={() => addOneMoreToCart(itemId, price)}>+</button> 
                 <button onClick={ () => deleteOneMoreFromCart(itemId, price, count) }>-</button> 
-               {(admin) &&
-               <>
-               <button onClick={() => setEditMode(true)}>Edit</button> 
-               <button onClick={() => handleDeleteItem()}>Delete</button> 
-               </>
-               }
-            </>) : 
-            (<EditListItem name={name} setEditMode={setEditMode} itemId={itemId} saveItemName={saveItemName}/>)}  
+            </>
+        )
+    }
+
+    if(editMode) {
+        editModeLayout = (
+             <>
+                <EditListItem name={name} setEditMode={setEditMode} itemId={itemId} saveItemName={saveItemName}/> 
+             </>
+        )
+
+    }
+    return ( 
+        <li>
+            {(!editMode) ? (defaultListLayout) : (editModeLayout)}
         </li>
         
      );
